# Universal-AI-Gateway/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Universal LLM Gateway",
    "description": "Universal LLM Gateway \u2014 unified API for multiple LLM providers",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/chat/completions": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Chat Completions",
        "description": "OpenAI-compatible chat completions endpoint (Requirement 1.1).\n\nIntegrates routing, token counting, cost calculation,\ncaching, and returns standardized response with required headers.",
        "operationId": "chat_completions_v1_chat_completions_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Unique request ID",
                "schema": {
                  "type": "string"
                }
              },
              "X-Provider": {
                "description": "Provider that handled the request",
                "schema": {
                  "type": "string"
                }
              },
              "X-Cache-Status": {
                "description": "HIT, MISS, or BYPASS",
                "schema": {
                  "type": "string"
                }
              },
              "X-Response-Time-Ms": {
                "description": "Gateway latency",
                "schema": {
                  "type": "string"
                }
              },
              "X-Correlation-ID": {
                "description": "Tracing ID",
                "schema": {
                  "type": "string"
                }
              },
              "X-Tenant-Budget-Remaining": {
                "description": "Remaining budget (optional)",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "402": {
            "description": "Budget Exceeded"
          },
          "403": {
            "description": "Security Policy Violation"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "description": "Model identifier (e.g. gpt-4o, claude-3-5-sonnet)"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array",
            "title": "Messages",
            "description": "Conversation messages"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature",
            "description": "Sampling temperature",
            "default": 0.7
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens",
            "description": "Maximum tokens to generate"
          },
          "top_p": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top P",
            "description": "Nucleus sampling"
          },
          "stream": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream",
            "description": "Enable streaming response",
            "default": false
          },
          "stop": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop",
            "description": "Stop sequences"
          },
          "presence_penalty": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": -2.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Presence Penalty",
            "default": 0.0
          },
          "frequency_penalty": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": -2.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Frequency Penalty",
            "default": 0.0
          },
          "routing_tags": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routing Tags",
            "description": "Custom routing tags"
          },
          "ensemble_strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ensemble Strategy",
            "description": "Strategy for model ensembling: 'fastest', 'longest'"
          },
          "ensemble_models": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ensemble Models",
            "description": "List of models to ensemble against"
          }
        },
        "type": "object",
        "required": [
          "model",
          "messages"
        ],
        "title": "ChatRequest",
        "description": "OpenAI-compatible chat completion request.\nSupports Requirement 1.5: follow OpenAI request schema."
      },
      "ChatResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Response ID"
          },
          "object": {
            "type": "string",
            "title": "Object",
            "description": "Object type",
            "default": "chat.completion"
          },
          "created": {
            "type": "integer",
            "title": "Created",
            "description": "Unix timestamp"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "Model used"
          },
          "choices": {
            "items": {
              "$ref": "#/components/schemas/Choice"
            },
            "type": "array",
            "title": "Choices",
            "description": "Completion choices"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage",
            "description": "Token usage"
          },
          "gateway_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GatewayMetadata"
              },
              {
                "type": "null"
              }
            ],
            "description": "Gateway metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created",
          "model",
          "choices"
        ],
        "title": "ChatResponse",
        "description": "OpenAI-compatible chat completion response.\nSupports Requirement 1.3: standardized response format."
      },
      "Choice": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index",
            "default": 0
          },
          "message": {
            "$ref": "#/components/schemas/Message"
          },
          "finish_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish Reason",
            "description": "Reason generation stopped",
            "default": "stop"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Choice",
        "description": "A single completion choice."
      },
      "GatewayMetadata": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "Provider that handled the request"
          },
          "cache_status": {
            "type": "string",
            "title": "Cache Status",
            "description": "HIT, MISS, or BYPASS",
            "default": "MISS"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms",
            "description": "Total gateway latency in milliseconds",
            "default": 0
          },
          "cost_usd": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost Usd",
            "description": "Estimated cost in USD",
            "default": "0.0"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Unique request correlation ID"
          },
          "routing_decision": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routing Decision",
            "description": "Routing details"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "request_id"
        ],
        "title": "GatewayMetadata",
        "description": "Gateway-specific metadata attached to every response.\nSupports Requirements 4.5 (routing metadata), 5.5 (cache status), 9.5 (cost)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Role: system, user, assistant, or tool"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Message content"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Optional name for the participant"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "Message",
        "description": "A single message in the conversation."
      },
      "Usage": {
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "title": "Prompt Tokens",
            "default": 0
          },
          "completion_tokens": {
            "type": "integer",
            "title": "Completion Tokens",
            "default": 0
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens",
            "default": 0
          }
        },
        "type": "object",
        "title": "Usage",
        "description": "Token usage statistics."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}